:tada:controllers, models and likes

+++ fellow-app/app/Http/Controllers/Posts/PostsController.php
@@ -6,11 +6,16 @@ use App\Http\Controllers\Controller;
 use App\Services\PostsService;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
-
+use App\Helpers\Helper;
 class PostsController extends Controller
 {
-    public function __construct(PostsService $postsService)
+    protected $postsService;
+    protected $helper;
+
+    public function __construct(PostsService $postsService, Helper $helper)
     {
+        $this->user = Auth()->guard('api')->user();
+        $this->helper = $helper;
         $this->postsService = $postsService;
     }
     /**
@@ -19,10 +24,10 @@ class PostsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {
+    {   
         $posts = $this->postsService->findAllPosts();
 
-        return response()->json($posts);
+        return response()->json($posts); 
     }
 
     /**
@@ -35,6 +40,14 @@ class PostsController extends Controller
     {
         $data = $request->all();
 
+        $user = $this->user;
+
+        $authVerify = $this->helper->verifyAuthUser($user->id);
+        if(!$authVerify) return response()->json(["msg" => "Erro encontrado! Usário inexistente!"]);
+
+        $authentic = $this->helper->authenticUser($user->id, $data['user_id']);
+        if(!$authentic) return response()->json(["msg" => "Você não é autor do post!"]);
+
         /* Aplicar contracts */
         $validator = Validator::make($data, [
             "type" => "numeric|min:0|max:3",
@@ -43,7 +56,7 @@ class PostsController extends Controller
         ]);
 
         if($validator->fails()) return response()->json([ 'errors' => $validator->errors() ], 400);
-
+        
         $newPost = $this->postsService->storePost($data);
 
         return response()->json($newPost);
@@ -71,7 +84,14 @@ class PostsController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $user = $this->user;
         $data = $request->all();
+
+        $authVerify = $this->helper->verifyAuthUser($user->id);
+        if(!$authVerify) return response()->json(["msg" => "Erro encontrado! Usário inexistente!"]);
+
+        $authentic = $this->helper->authenticUser($user->id, $data['user_id']);
+        if(!$authentic) return response()->json(["msg" => "Você não é autor do post!"]);
 
         $postUpdate = $this->postsService->updatePost($data, $id);
 
@@ -86,6 +106,13 @@ class PostsController extends Controller
      */
     public function destroy($id)
     {
+        $user = $this->user;
+
+        $authVerify = $this->helper->verifyAuthUser($user->id);
+        if(!$authVerify) return response()->json(["msg" => "Erro encontrado! Usário inexistente!"]);
+        
+        /* não esquecer de verificar autenticidade */
+
         $deletedPost = $this->postsService->deletePost($id);
 
         return response()->json($deletedPost);
